Make stable_id utilities separate and private
