+++ src/fonduer/candidates/models/implicitspan.py
@@ -5,7 +5,8 @@ from sqlalchemy.sql import text
 from sqlalchemy.types import PickleType
 
 from fonduer.candidates.models.span import TemporarySpan
-from fonduer.parser.models.context import Context, split_stable_id
+from fonduer.parser.models.context import Context
+from fonduer.parser.models.utils import split_stable_id
 
 
 class TemporaryImplicitSpan(TemporarySpan):
