+++ src/fonduer/candidates/models/span.py
@@ -3,7 +3,8 @@ from sqlalchemy.orm import backref, relationship
 from sqlalchemy.types import PickleType
 
 from fonduer.candidates.models.temporarycontext import TemporaryContext
-from fonduer.parser.models.context import Context, construct_stable_id
+from fonduer.parser.models.context import Context
+from fonduer.parser.models.utils import construct_stable_id
 
 
 class TemporarySpan(TemporaryContext):
