+++ src/fonduer/parser/models/context.py
@@ -13,48 +13,15 @@ class Context(_meta.Base):
 
     __tablename__ = "context"
 
-    #: The unique id of the Context.
+    #: The unique id of the ``Context``.
     id = Column(Integer, primary_key=True)
 
-    #: The type of the context represented as a String (e.g. "sentence",
+    #: The type of the ``Context`` represented as a string (e.g. "sentence",
     #: "paragraph", "figure").
     type = Column(String, nullable=False)
 
-    #: A stable representation of the Context that will not change between
+    #: A stable representation of the ``Context`` that will not change between
     #: runs.
     stable_id = Column(String, unique=True, nullable=False)
 
     __mapper_args__ = {"polymorphic_identity": "context", "polymorphic_on": type}
-
-
-def construct_stable_id(
-    parent_context,
-    polymorphic_type,
-    relative_char_offset_start,
-    relative_char_offset_end,
-):
-    """
-    Contruct a stable ID for a Context given its parent and its character
-    offsets relative to the parent.
-    """
-    doc_id, _, parent_doc_char_start, _ = split_stable_id(parent_context.stable_id)
-    start = parent_doc_char_start + relative_char_offset_start
-    end = parent_doc_char_start + relative_char_offset_end
-    return "{}::{}:{}:{}".format(doc_id, polymorphic_type, start, end)
-
-
-def split_stable_id(stable_id):
-    """Split stable id, returning:
-
-        * Document (root) stable ID
-        * Context polymorphic type
-        * Character offset start, end *relative to document start*
-
-    Returns tuple of four values.
-    """
-    split1 = stable_id.split("::")
-    if len(split1) == 2:
-        split2 = split1[1].split(":")
-        if len(split2) == 3:
-            return split1[0], split2[0], int(split2[1]), int(split2[2])
-    raise ValueError("Malformed stable_id:", stable_id)
